modifiedd

+++ src/main/java/com/website/FoodOrder/model/Order.java
@@ -41,7 +41,7 @@ public class Order {
 
     private Integer totalItem;
 
-    private Integer totalPrice;
+    private Long totalPrice;
